Image loading by loader thread

+++ src/renderlib/textures/raw_image.rs
@@ -6,6 +6,6 @@ pub struct RawImage {
     pub height: u32,
     pub format: vk::Format,
     pub pixels: Vec<u8>,
-    pub size_bytes: u32,
+    pub size_bytes: u64,
     pub mip_levels: u32,
 }
